requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pypdf2==3.0.1
swarmauri==0.4.0
python-dotenv==1.0.1